@@ -1,6 +1,6 @@
 name: connectivity_plus
 description: Flutter plugin for discovering the state of the network (WiFi & mobile/cellular) connectivity on Android and iOS.
-version: 0.8.0
+version: 0.9.0-nullsafety.0
 homepage: https://plus.fluttercommunity.dev/
 repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
 
@@ -24,12 +24,12 @@ flutter:
 dependencies:
   flutter:
     sdk: flutter
-  meta: ^1.0.5
-  connectivity_plus_platform_interface: ^0.4.0
-  connectivity_plus_linux: ^0.3.0
-  connectivity_plus_macos: ^0.4.0
-  connectivity_plus_web: ^0.6.0
-  connectivity_plus_windows: ^0.1.0
+  meta: ^1.3.0-nullsafety.6
+  connectivity_plus_platform_interface: ^0.5.0-nullsafety.0
+  connectivity_plus_linux: ^0.4.0-nullsafety.0
+  connectivity_plus_macos: ^0.5.0-nullsafety.0
+  connectivity_plus_web: ^0.7.0-nullsafety.0
+  connectivity_plus_windows: ^0.2.0-nullsafety.0
 
 dev_dependencies:
   flutter_test:
@@ -37,10 +37,13 @@ dev_dependencies:
   flutter_driver:
     sdk: flutter
   test: any
-  mockito: ^4.1.1
-  plugin_platform_interface: ^1.0.0
-  pedantic: ^1.9.2
+  mockito: ^5.0.0-nullsafety.7
+  plugin_platform_interface: ^1.1.0-nullsafety.2
+  pedantic: ^1.10.0-nullsafety.3
+
+dependency_overrides:
+  args: '^2.0.0-nullsafety'
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.12.13+hotfix.5"
